fdt: Add Kaen keyboard support
This adds support for the Kaen keyboard, and fixes a few things wrong in the
keyboard implementation.

Note that there is a problem that the Shift key generates a DF code, which it
should not. This doesn't happen on Seaboard and Aebl.

BUG=chromium-os:17034
TEST=boot U-Boot; type plain, shift and fn keys and check that the correct
output results

Reviewed-on: http://gerrit.chromium.org/gerrit/3414
Tested-by: Simon Glass <sjg@chromium.org>

+++ include/tegra-kbc.h
@@ -27,10 +27,10 @@ enum {
 
 struct tegra_keyboard_config {
 	/* keycode tables, one for each row/col position */
-	int plain_keycode[KBC_KEY_COUNT]; /* when no Shift or Fn */
-	int shift_keycode[KBC_KEY_COUNT]; /* Shift modifier key is pressed */
-	int fn_keycode[KBC_KEY_COUNT]; /* Fn modifier key is pressed */
-	int ctrl_keycode[KBC_KEY_COUNT]; /* Ctrl modifier key is pressed */
+	u8 plain_keycode[KBC_KEY_COUNT]; /* when no Shift or Fn */
+	u8 shift_keycode[KBC_KEY_COUNT]; /* Shift modifier key is pressed */
+	u8 fn_keycode[KBC_KEY_COUNT]; /* Fn modifier key is pressed */
+	u8 ctrl_keycode[KBC_KEY_COUNT]; /* Ctrl modifier key is pressed */
 };
 
 struct kbc_tegra {
